Relax setindex argument types

+++ src/ArrayInterface.jl
@@ -19,7 +19,7 @@ ismutable(::Type{<:Array}) = true
 ismutable(::Type{<:Number}) = false
 
 # Piracy
-function Base.setindex(x::AbstractArray,v,i::Int...)
+function Base.setindex(x::AbstractArray,v,i...)
   _x = copy(x)
   _x[i...] = v
   _x
